+++ src/cpp_echo/src/class/server.cpp
@@ -1,36 +1,50 @@
 class Server : public rclcpp::Node{
 public:
-  Server() : Node("server")
-  {
-	//declare parameter
-	this->declare_parameter("number_publisher", rclcpp::PARAMETER_INTEGER); 
-	
-	//define callback group to implement mutual exclusion
-	auto my_callback_group = create_callback_group(rclcpp::CallbackGroupType::MutuallyExclusive);
-	rclcpp::SubscriptionOptions options;
-	options.callback_group = my_callback_group;
-	
-	//create n pub/sub, one for each publisher
-    for(int i = 0; i < this->get_parameter("number_publisher").as_int(); i++)
+    Server() : Node("server")
     {
-      //convert number into alphabetic char
-      std::string lettera(1, intToAlphabet(i+1));
-      
-      std::string topic_name = "topic_" + lettera;
-      std::string echo_name = "echo_" + lettera;
-      
-      publisher_.push_back(this->create_publisher<std_msgs::msg::String>(
-        echo_name, 100));
+    /***
+     * define Server Node
+     * 
+     * Args:
+     * 		number_publisher ([int]): number of current client
+    */
+    //declare parameter
 
-      subscription_.push_back(this->create_subscription<std_msgs::msg::String>(
-        topic_name, 100, std::bind(&Server::topic_callback, this, _1), options));
+    //declare parameter
+    this->declare_parameter("number_publisher", rclcpp::PARAMETER_INTEGER); 
+    
+    //define callback group to implement mutual exclusion
+    auto my_callback_group = create_callback_group(rclcpp::CallbackGroupType::MutuallyExclusive);
+    rclcpp::SubscriptionOptions options;
+    options.callback_group = my_callback_group;
+    
+    //create n pub/sub, one for each publisher
+      for(int i = 0; i < this->get_parameter("number_publisher").as_int(); i++)
+      {
+        //convert number into alphabetic char
+        std::string lettera(1, intToAlphabet(i+1));
+        
+        std::string topic_name = "topic_" + lettera;
+        std::string echo_name = "echo_" + lettera;
+        
+        publisher_.push_back(this->create_publisher<std_msgs::msg::String>(
+          echo_name, 100));
+
+        subscription_.push_back(this->create_subscription<std_msgs::msg::String>(
+          topic_name, 100, std::bind(&Server::topic_callback, this, _1), options));
+      }
     }
-  }
 
 private:
   void topic_callback(const std_msgs::msg::String::SharedPtr msg) const
   {
-	//read message from a subscription
+    /***
+     * read message from server
+     * 
+     * Args:
+     * 		msg ([std_msgs]): message rom server node
+    */
+    //read message from a subscription
     RCLCPP_INFO(this->get_logger(), "I heard: '%s'", msg->data.c_str());
     
     //exctract which subscriptions it is
