+++ src/cpp_echo/src/utils/utils.cpp
@@ -2,17 +2,35 @@
 #include <iostream>
 #include <fstream>
 
-//convert int number into alphabetic char
 char intToAlphabet( int i )
 {
-   return static_cast<char>('A' - 1 + i);
+	/***
+	 * convert int number into alphabetic char
+	 * 
+	 * Args:
+	 * 		i ([int]): number to convert into char
+	 * 
+	 * Return:
+	 * 		char alphabetic letter
+	*/
+	return static_cast<char>('A' - 1 + i);
 }
 
-//save result on file
 int save_to_file(std::string filename_, 
 				 std::string data_)
 {
-	//check if file exists, and if it will not, we create it
+	/***
+	 * save result on file a specific file
+	 * 
+	 * Args:
+	 * 		filename_ ([std::string]): name of file into save result
+	 * 		data_ ([std::string]): result from a specific message recived back from server
+	 * 
+	 * Return:
+	 * 		0 if there isn't error
+	 * 		1 is file can't be open
+	*/
+	
 	std::ofstream writer("result/" + filename_, std::ios::app);
 
 	//check if we can write on it
@@ -35,7 +53,19 @@ std::string concatenete_result( int publisher_,
 								int end_, 
 								int start_)
 {
-	//concate result and return it
+	/***
+	 * create the result string 
+	 * 
+	 * Args:
+	 * 		publisher_ ([int]): number of current client
+	 * 		publisher_max ([int]): total number of client
+	 * 		message_size_ ([int]): size of a single messagge
+	 * 		end_ ([int]): arrival time of message
+	 * 		start_ ([int]): send time of message
+	 * 
+	 * Return:
+	 * 		string contained all the parameter
+	*/
 	std::string result_ = 
 	
 	return "client" + std::to_string(publisher_) + "-"
